Try extracting copied course
